+++ AddOns/Platynator/Core/Migration.lua
@@ -679,6 +679,32 @@ local function UpgradeDesignv16(design)
   end
 end
 
+local function UpgradeDesignv17(design)
+  for _, auras in ipairs(design.auras) do
+    if auras.kind == "buffs" then
+      auras.filters.friendlyFromYou = true
+    end
+  end
+end
+
+local function UpgradeDesignv18(design)
+  for _, auras in ipairs(design.auras) do
+    if auras.kind == "buffs" then
+      auras.filters.dispelable = nil
+      auras.filters.dispellable = true
+    end
+  end
+end
+
+local function UpgradeDesignv19(design)
+  for _, bar in ipairs(design.specialBars) do
+    if bar.kind == "power" then
+      bar.useSpecColors = bar.useSpecColors ~= nil and bar.useSpecColors or true
+      bar.fixedColor = bar.fixedColor or GetColor("f0c900")
+    end
+  end
+end
+
 local designUpgrades = {
   UpgradeDesignv1,
   UpgradeDesignv2,
@@ -696,6 +722,9 @@ local designUpgrades = {
   UpgradeDesignv14,
   UpgradeDesignv15,
   UpgradeDesignv16,
+  UpgradeDesignv17,
+  UpgradeDesignv18,
+  UpgradeDesignv19,
 }
 
 function addonTable.Core.UpgradeDesign(design)
@@ -845,6 +874,21 @@ local function MigrateSettingsv6()
   -- Removed as this was the wrong place
 end
 
+local function MigrateSettingsv7()
+  local filters = addonTable.Config.Get(addonTable.Config.Options.AURA_FILTERS)
+  local globalCrowdControl = {include = {}, exclude = {}}
+
+  for key, val in pairs(filters) do
+    if val.crowdControl then
+      Mixin(globalCrowdControl.include, val.crowdControl.include)
+      Mixin(globalCrowdControl.exclude, val.crowdControl.exclude)
+      val.crowdControl = nil
+    end
+  end
+
+  filters.crowdControl = globalCrowdControl
+end
+
 local settingUpgrades = {
   MigrateSettingsv1,
   MigrateSettingsv2,
@@ -853,6 +897,7 @@ local settingUpgrades = {
   MigrateSettingsv5,
   MigrateSettingsv6,
   MigrateSettingsv6,
+  MigrateSettingsv7,
 }
 function addonTable.Core.MigrateSettings()
   if #settingUpgrades + 1 ~= addonTable.Config.Get(addonTable.Config.Options.MIGRATION) then