+++ AddOns/Platynator/CustomiseDialog/AddWidgetsList.lua
@@ -157,6 +157,8 @@ addonTable.CustomiseDialog.DesignWidgets = {
       scale = 1 * 0.60,
       layer = 3,
       anchor = {"TOPLEFT", -140, 50},
+      useSpecColors = true,
+      fixedColor = GetColor("f0c900"),
     },
   },
   {
@@ -338,9 +340,11 @@ addonTable.CustomiseDialog.DesignWidgets = {
         reversed = false,
       },
       filters = {
-        dispelable = false,
+        dispellable = true,
         important = true,
         enrage = false,
+        defensive = false,
+        friendlyFromYou = true,
       },
       showType = true,
       showSwipe = true,